+++ src/Service/TimestreamWrite/src/TimestreamWriteClient.php
@@ -14,12 +14,41 @@ use AsyncAws\TimestreamWrite\Exception\RejectedRecordsException;
 use AsyncAws\TimestreamWrite\Exception\ResourceNotFoundException;
 use AsyncAws\TimestreamWrite\Exception\ThrottlingException;
 use AsyncAws\TimestreamWrite\Exception\ValidationException;
+use AsyncAws\TimestreamWrite\Input\DescribeEndpointsRequest;
 use AsyncAws\TimestreamWrite\Input\WriteRecordsRequest;
+use AsyncAws\TimestreamWrite\Result\DescribeEndpointsResponse;
 use AsyncAws\TimestreamWrite\Result\WriteRecordsResponse;
 use AsyncAws\TimestreamWrite\ValueObject\Record;
 
 class TimestreamWriteClient extends AbstractApi
 {
+    /**
+     * DescribeEndpoints returns a list of available endpoints to make Timestream API calls against. This API is available
+     * through both Write and Query.
+     *
+     * @see https://docs.aws.amazon.com/timestream/latest/developerguide/API_DescribeEndpoints.html
+     * @see https://docs.aws.amazon.com/aws-sdk-php/v3/api/api-ingest.timestream-2018-11-01.html#describeendpoints
+     *
+     * @param array{
+     *   @region?: string,
+     * }|DescribeEndpointsRequest $input
+     *
+     * @throws InternalServerException
+     * @throws ValidationException
+     * @throws ThrottlingException
+     */
+    public function describeEndpoints($input = []): DescribeEndpointsResponse
+    {
+        $input = DescribeEndpointsRequest::create($input);
+        $response = $this->getResponse($input->request(), new RequestContext(['operation' => 'DescribeEndpoints', 'region' => $input->getRegion(), 'exceptionMapping' => [
+            'InternalServerException' => InternalServerException::class,
+            'ValidationException' => ValidationException::class,
+            'ThrottlingException' => ThrottlingException::class,
+        ]]));
+
+        return new DescribeEndpointsResponse($response);
+    }
+
     /**
      * The WriteRecords operation enables you to write your time series data into Timestream. You can specify a single data
      * point or a batch of data points to be inserted into the system. Timestream offers you with a flexible schema that
@@ -61,9 +90,14 @@ class TimestreamWriteClient extends AbstractApi
             'AccessDeniedException' => AccessDeniedException::class,
             'RejectedRecordsException' => RejectedRecordsException::class,
             'InvalidEndpointException' => InvalidEndpointException::class,
-        ]]));
+        ], 'requiresEndpointDiscovery' => true, 'usesEndpointDiscovery' => true]));
 
         return new WriteRecordsResponse($response);
+    }
+
+    protected function discoverEndpoints(?string $region): array
+    {
+        return $this->describeEndpoints($region ? ['@region' => $region] : [])->getEndpoints();
     }
 
     protected function getAwsErrorFactory(): AwsErrorFactoryInterface